Add adaptive feedback-lag predictor to active control
Steering against a stale grid meter (transport/measurement delay plus the
hold between refreshes) is the dominant source of overshoot and hunting:
the loop keeps reacting to a grid that doesn't yet reflect the output it
already commanded. Rather than encoding per-firmware steering quirks or
making users tune a throttle interval, advance the stale reading to the
present using the batteries' own fresh reported outputs:

    predicted = grid_stale - (Σout_now - Σout_{now - L})

The effective lag L is learned online per install: for each candidate lag
track the variance of the first difference of (grid_stale + Σout_lagged);
the minimiser is the delay at which the battery's own output ripple
cancels out, leaving smooth household consumption. A fast meter has no
ripple to cancel at any nonzero lag, so it self-limits to ~0 compensation.

Across the steering-evaluation matrix this improves every metric: mean
overshoot -58%, worst-case overshoot -43%, band crossings -50%, battery
travel -24%, settle p95 -32%, with low-latency single-battery scenarios
unchanged (overall geomean 0.75). On by default; bound/disable with the
new PREDICT_LAG_S option (predict_lag_s on ESPHome).

Ported to the C++ balancer in lockstep. While wiring up the differential
parity suite this surfaced and fixed a real pre-existing port bug: the C++
efficiency cache stored the pre-swap priority key after a saturation swap
(Python recomputes it post-swap), so the next tick missed the cache and
diverged. Also bumped the C++ saturation/fade EMAs and the swap threshold
to double to match Python's canonical math and stop float drift flipping
discrete decisions on a knife-edge.

+++ esphome/components/ct002/ct002.h
@@ -115,7 +115,7 @@ class CT002Component : public Component {
   // Balancer configuration setter (called once from to_code() after
   // populating a BalancerConfig).
   void set_balancer_config(const BalancerConfig &cfg) { this->balancer_cfg_ = cfg; }
-  void set_balancer_saturation(float alpha, float min_target, float decay_factor,
+  void set_balancer_saturation(double alpha, float min_target, double decay_factor,
                               float grace_seconds, float stall_timeout_seconds, bool enabled) {
     this->saturation_alpha_ = alpha;
     this->saturation_min_target_ = min_target;
@@ -297,9 +297,10 @@ class CT002Component : public Component {
   // MQTT-insights poll_interval matches between stacks.
   static constexpr float POLL_INTERVAL_EMA_ALPHA = 0.3f;
   BalancerConfig balancer_cfg_;
-  float saturation_alpha_{0.15f};
+  // double (matching Python) so the saturation EMA bit-matches across stacks.
+  double saturation_alpha_{0.15};
   float saturation_min_target_{20.0f};
-  float saturation_decay_factor_{0.995f};
+  double saturation_decay_factor_{0.995};
   float saturation_grace_seconds_{90.0f};
   float saturation_stall_timeout_seconds_{60.0f};
   bool saturation_enabled_{true};
